refactor => HalfEdgeGraph2d.cs
